Create root component with all providers. Change some styling.
