+++ src/components/Header.tsx
@@ -1,10 +1,18 @@
+import { H1 } from 'native-base'
 import React from 'react'
-import { Text, View } from 'react-native'
-import { styles } from '../styles'
+import { StyleSheet, View } from 'react-native'
 
 export const Header = () => (
-  <View style={styles.sectionMainTitle}>
-    <Text style={styles.mainTitle}>PUC the</Text>
-    <Text style={styles.mainTitle}>Pull-Up-Counter</Text>
+  <View style={styles.mainTitle}>
+    <H1>PUC the</H1>
+    <H1>Pull-Up-Counter</H1>
   </View>
 )
+const styles = StyleSheet.create({
+  mainTitle: {
+    alignItems: 'center',
+    justifyContent: 'center',
+    marginVertical: 32,
+    paddingHorizontal: 24,
+  },
+})
